pyproject.toml
[project]
name = "pdf-stories"
version = "0.1.0"
description = "Play PPT Interactive Stories"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-blob>=12.24.0",
    "flask>=3.1.1",
    "flask-admin>=1.6.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "jinja-layout>=0.4.0",
    "pysondb>=1.6.7",
    "python-dotenv>=1.1.1",
    "python-pptx>=1.0.2",
    "watchdog>=6.0.0",
    "wtforms==3.1.2",
]
